Se agrega guardado parcial para formato de audiencia Se agregan reglas para funcionalidad de radios y para posteriores formatos de audiencia
PROBAR FUNCIONALIDAD

+++ src/Umeca/src/main/java/com/umeca/service/supervisor/HearingFormatServiceImpl.java
@@ -246,10 +246,12 @@ public class HearingFormatServiceImpl implements HearingFormatService {
                     List<AssignedArrangement> oldArrangements = hearingFormat.getAssignedArrangements();
                     hearingFormat.setAssignedArrangements(null);
 
-                    for (AssignedArrangement actArr : oldArrangements) {
-                        actArr.setHearingFormat(null);
-                        actArr.setArrangement(null);
-                        assignedArrangementRepository.delete(actArr);
+                    if (oldArrangements != null) {
+                        for (AssignedArrangement actArr : oldArrangements) {
+                            actArr.setHearingFormat(null);
+                            actArr.setArrangement(null);
+                            assignedArrangementRepository.delete(actArr);
+                        }
                     }
 
                     hearingFormat.setAssignedArrangements(lstNewAssigArrmnt);
@@ -282,10 +284,13 @@ public class HearingFormatServiceImpl implements HearingFormatService {
                     List<ContactData> oldContacts = hearingFormat.getContacts();
                     hearingFormat.setContacts(null);
 
-                    for (ContactData act : oldContacts) {
-                        act.setHearingFormat(null);
-                        contactDataRepository.delete(act);
+                    if (oldContacts != null) {
+                        for (ContactData act : oldContacts) {
+                            act.setHearingFormat(null);
+                            contactDataRepository.delete(act);
+                        }
                     }
+
 
                     hearingFormat.setContacts(lstNewContactData);
                 } else {
@@ -329,6 +334,7 @@ public class HearingFormatServiceImpl implements HearingFormatService {
             hearingFormatView.setCanSave(true);
             hearingFormatView.setCanEdit(true);
             hearingFormatView.setDisableAll(false);
+            hearingFormatView.setHasPrevHF(true);
 
         } else {//si no existe un formato de audiencia anterior
             //evaluo el origen del meeting
@@ -349,6 +355,7 @@ public class HearingFormatServiceImpl implements HearingFormatService {
                 hearingFormatView.setCanSave(true);
                 hearingFormatView.setCanEdit(true);
                 hearingFormatView.setDisableAll(false);
+                hearingFormatView.setHasPrevHF(false);
 
             } else if (meetType.equals(HearingFormatConstants.MEETING_PROCEDURAL_RISK)) {//si el meeting fue creado normalmente y ya esta completo
 
@@ -380,6 +387,7 @@ public class HearingFormatServiceImpl implements HearingFormatService {
                 hearingFormatView.setCanSave(true);
                 hearingFormatView.setCanEdit(true);
                 hearingFormatView.setDisableAll(false);
+                hearingFormatView.setHasPrevHF(false);
 
             }
         }
@@ -616,7 +624,7 @@ public class HearingFormatServiceImpl implements HearingFormatService {
                 hearingFormat.getCaseDetention().setStatus(statusCaseRepository.findByCode(Constants.CASE_STATUS_HEARING_FORMAT_INCOMPLETE));
             }
 
-            if (hearingFormat.getHearingFormatSpecs() != null && hearingFormat.getHearingFormatSpecs().getLinkageProcess() != null &&
+            if (hearingFormat.getIsFinished() != null && hearingFormat.getIsFinished() == true && hearingFormat.getHearingFormatSpecs() != null && hearingFormat.getHearingFormatSpecs().getLinkageProcess() != null &&
                     hearingFormat.getHearingFormatSpecs().getLinkageProcess().equals(HearingFormatConstants.PROCESS_VINC_NO)) {
 
                 hearingFormat.getCaseDetention().setStatus(statusCaseRepository.findByCode(Constants.CASE_STATUS_PRE_CLOSED));
@@ -632,17 +640,20 @@ public class HearingFormatServiceImpl implements HearingFormatService {
 
             hearingFormat = hearingFormatRepository.save(hearingFormat);
 
-            sb = new StringBuilder();
-            sb.append(request.getContextPath());
+            response.setHasError(false);
 
             if (hearingFormat.getIsFinished() == true) {
+                sb = new StringBuilder();
+                sb.append(request.getContextPath());
                 sb.append("/supervisor/hearingFormat/indexFormats.html?id=");
                 sb.append(hearingFormat.getCaseDetention().getId());
+                response.setUrlToGo(sb.toString());
+            } else {
+                response.setMessage(hearingFormat.getId() + "|Se ha registrado el formato de audiencia.");
+
             }
 
-            response.setHasError(false);
-            response.setMessage(hearingFormat.getId() + "|Se ha registrado el formato de audiencia.");
-            response.setUrlToGo(sb.toString());
+
         } catch (Exception e) {
             System.out.println("Ha ocurrido un error al guardar el formato de audiencia (serviceImpl)!!!");
             e.printStackTrace();
